rename/#316: 폴더링 수정
