+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Data/Providers/AppleLoginProviderImpl.swift
@@ -1,7 +1,7 @@
+import AuthenticationServices
 import UIKit
 
 import MLSAuthFeatureInterface
-import AuthenticationServices
 
 import RxSwift
