+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Domain/Providers/SocialAuthenticatableProvider.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 import MLSAuthFeatureInterface
 
 import RxSwift
