+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Domain/Repositories/TokenRepository.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 public protocol TokenRepository {
     func fetchToken(type: TokenType) -> Result<String, Error>
     func saveToken(type: TokenType, value: String) -> Result<Void, Error>
