+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Domain/UseCases/AuthAPI/SignUpWithAppleUseCaseImpl.swift
@@ -1,10 +1,8 @@
-import Foundation
-
 import MLSAuthFeatureInterface
 
 import RxSwift
 
-public final class SignUpWithKakaoUseCaseImpl: SignUpWithKakaoUseCase {
+public final class SignUpWithAppleUseCaseImpl: SignUpWithAppleUseCase {
     private let authRepository: AuthAPIRepository
     private let tokenRepository: TokenRepository
     private let userDefaultsRepository: UserDefaultsRepository
@@ -20,11 +18,11 @@ public final class SignUpWithKakaoUseCaseImpl: SignUpWithKakaoUseCase {
     }
 
     public func execute(credential: Credential, isMarketingAgreement: Bool, fcmToken: String?) -> Observable<SignUpResponse> {
-        return authRepository.signUpWithKakao(credential: credential, isMarketingAgreement: isMarketingAgreement, fcmToken: fcmToken)
+        return authRepository.signUpWithApple(credential: credential, isMarketingAgreement: isMarketingAgreement, fcmToken: fcmToken)
             .flatMap { response -> Observable<SignUpResponse> in
                 let saveAccess = self.tokenRepository.saveToken(type: .accessToken, value: response.accessToken)
                 let saveRefresh = self.tokenRepository.saveToken(type: .refreshToken, value: response.refreshToken)
-                let savePlatform = self.userDefaultsRepository.savePlatform(platform: .kakao)
+                let savePlatform = self.userDefaultsRepository.savePlatform(platform: .apple)
 
                 switch (saveAccess, saveRefresh) {
                 case (.success, .success):
