+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Domain/UseCases/AuthAPI/SignUpWithKakaoUseCaseImpl.swift
@@ -1,10 +1,8 @@
-import Foundation
-
 import MLSAuthFeatureInterface
 
 import RxSwift
 
-public class LoginWithKakaoUseCaseImpl: LoginWithKakaoUseCase {
+public final class SignUpWithKakaoUseCaseImpl: SignUpWithKakaoUseCase {
     private let authRepository: AuthAPIRepository
     private let tokenRepository: TokenRepository
     private let userDefaultsRepository: UserDefaultsRepository
@@ -19,32 +17,19 @@ public class LoginWithKakaoUseCaseImpl: LoginWithKakaoUseCase {
         self.userDefaultsRepository = userDefaultsRepository
     }
 
-    public func execute(credential: Credential) -> Observable<LoginResponse> {
-        return authRepository.loginWithKakao(credential: credential)
-            .flatMap { response -> Observable<LoginResponse> in
+    public func execute(credential: Credential, isMarketingAgreement: Bool, fcmToken: String?) -> Observable<SignUpResponse> {
+        return authRepository.signUpWithKakao(credential: credential, isMarketingAgreement: isMarketingAgreement, fcmToken: fcmToken)
+            .flatMap { response -> Observable<SignUpResponse> in
                 let saveAccess = self.tokenRepository.saveToken(type: .accessToken, value: response.accessToken)
                 let saveRefresh = self.tokenRepository.saveToken(type: .refreshToken, value: response.refreshToken)
                 let savePlatform = self.userDefaultsRepository.savePlatform(platform: .kakao)
 
-                guard case (.success, .success) = (saveAccess, saveRefresh) else {
+                switch (saveAccess, saveRefresh) {
+                case (.success, .success):
+                    return savePlatform.andThen(Observable.just(response))
+                default:
                     return Observable.error(TokenRepositoryError.dataConversionError(message: "Failed to save tokens"))
                 }
-
-                var fcmToken: String?
-                if case .success(let token) = self.tokenRepository.fetchToken(type: .fcmToken) {
-                    fcmToken = token
-                }
-
-                let fcmUpdate = if let fcmToken {
-                    self.authRepository.fcmToken(fcmToken: fcmToken)
-                        .catch { error in
-                            print("FCM token update failed: \(error)")
-                            return .empty()
-                        }
-                } else {
-                    Completable.empty()
-                }
-                return fcmUpdate.andThen(savePlatform).andThen(Observable.just(response))
             }
             .catch { error in
                 Observable.error(error)
