+++ MLS/MLSAuthFeature/Sources/MLSAuthFeature/Presentation/Login/LoginReactor.swift
@@ -1,8 +1,8 @@
 import UserNotifications
 
 import MLSAuthFeatureInterface
-import ReactorKit
 
+import ReactorKit
 import RxSwift
 
 public final class LoginReactor: Reactor {
